city/county join
Updated script so that city and county data is not brought into the fold until after the transformations are done on the political data. Also commented out the steps that create buffers for cities and counties, as we aren't going to use that anymore since we're doing a 1-to-1 join.